+++ jupyter_dash/jupyter_app.py
@@ -1,3 +1,5 @@
+import logging
+
 import dash
 import os
 import requests
@@ -9,6 +11,7 @@ import re
 import sys
 import inspect
 import traceback
+import threading
 import warnings
 import queue
 
@@ -18,8 +21,9 @@ from IPython.core.ultratb import FormattedTB
 from ansi2html import Ansi2HTMLConverter
 import uuid
 
+from werkzeug.serving import make_server
+
 from .comms import _dash_comm, _jupyter_config, _request_jupyter_config
-from ._stoppable_thread import StoppableThread
 
 
 def _get_skip(error: Exception):
@@ -50,7 +54,7 @@ class JupyterDash(dash.Dash):
     _in_colab = "google.colab" in sys.modules
     _token = str(uuid.uuid4())
 
-    _server_threads = {}
+    _servers = {}
 
     @classmethod
     def infer_jupyter_proxy_config(cls):
@@ -147,6 +151,7 @@ class JupyterDash(dash.Dash):
             return 'Alive'
 
         self.server.logger.disabled = True
+        self._exception_handling_added = False
 
     def run(
             self,
@@ -186,11 +191,8 @@ class JupyterDash(dash.Dash):
             return
 
         # Get host and port
-        host = kwargs.get("host", os.getenv("HOST", "127.0.0.1"))
-        port = kwargs.get("port", os.getenv("PORT", "8050"))
-
-        kwargs['host'] = host
-        kwargs['port'] = port
+        host = kwargs.pop("host", os.getenv("HOST", "127.0.0.1"))
+        port = int(kwargs.pop("port", os.getenv("PORT", "8050")))
 
         # Validate / infer display mode
         if JupyterDash._in_colab:
@@ -222,11 +224,10 @@ class JupyterDash(dash.Dash):
             inline_exceptions = mode == "inline"
 
         # Terminate any existing server using this port
-        old_server = self._server_threads.get((host, port))
+        old_server = self._servers.get((host, port))
         if old_server:
-            old_server.kill()
-            old_server.join()
-            del self._server_threads[(host, port)]
+            old_server.shutdown()
+            del self._servers[(host, port)]
 
         # Configure pathname prefix
         requests_pathname_prefix = self.config.get('requests_pathname_prefix', None)
@@ -256,11 +257,7 @@ class JupyterDash(dash.Dash):
         )
 
         # Default the global "debug" flag to True
-        debug = kwargs.get('debug', True)
-
-        # Disable debug flag when calling superclass because it doesn't work
-        # in notebook
-        kwargs['debug'] = False
+        debug = kwargs.pop('debug', True)
 
         # Enable supported dev tools
         if debug:
@@ -283,14 +280,32 @@ class JupyterDash(dash.Dash):
                 # there is no active kernel.
                 kwargs['dev_tools_hot_reload'] = mode == "external"
 
-        # suppress warning banner printed to standard out
-        flask.cli.show_server_banner = lambda *args, **kwargs: None
-
         # Set up custom callback exception handling
         self._config_callback_exception_handling(
             dev_tools_prune_errors=kwargs.get('dev_tools_prune_errors', True),
             inline_exceptions=inline_exceptions,
         )
+
+        dev_tools_args = dict(
+            debug=debug,
+            dev_tools_ui=kwargs.pop("dev_tools_ui", None),
+            dev_tools_props_check=kwargs.pop("dev_tools_props_check", None),
+            dev_tools_serve_dev_bundles=kwargs.pop("dev_tools_serve_dev_bundles", None),
+            dev_tools_hot_reload=kwargs.pop("dev_tools_hot_reload", None),
+            dev_tools_hot_reload_interval=kwargs.pop("dev_tools_hot_reload_interval", None),
+            dev_tools_hot_reload_watch_interval=kwargs.pop("dev_tools_hot_reload_watch_interval", None),
+            dev_tools_hot_reload_max_retry=kwargs.pop("dev_tools_hot_reload_max_retry", None),
+            dev_tools_silence_routes_logging=kwargs.pop("dev_tools_silence_routes_logging", None),
+            dev_tools_prune_errors=kwargs.pop("dev_tools_prune_errors", None),
+        )
+
+        if len(kwargs):
+            raise Exception(f"Invalid keyword argument: {list(kwargs.keys())}")
+
+        self.enable_dev_tools(**dev_tools_args)
+
+        # suppress warning banner printed to standard out
+        flask.cli.show_server_banner = lambda *args, **kw: None
 
         # prevent partial import of orjson when it's installed and mode=jupyterlab
         # TODO: why do we need this? Why only in this mode? Importing here in
@@ -302,6 +317,13 @@ class JupyterDash(dash.Dash):
 
         err_q = queue.Queue()
 
+        server = make_server(
+            host, port, self.server,
+            threaded=True,
+            processes=0
+        )
+        logging.getLogger("werkzeug").setLevel(logging.ERROR)
+
         @retry(
             stop_max_attempt_number=15,
             wait_exponential_multiplier=100,
@@ -309,18 +331,18 @@ class JupyterDash(dash.Dash):
         )
         def run():
             try:
-                super_run_server(**kwargs)
+                server.serve_forever()
             except SystemExit:
                 pass
             except Exception as error:
                 err_q.put(error)
                 raise error
 
-        thread = StoppableThread(target=run)
-        thread.setDaemon(True)
+        thread = threading.Thread(target=run)
+        thread.daemon = True
         thread.start()
 
-        self._server_threads[(host, port)] = thread
+        self._servers[(host, port)] = server
 
         # Wait for server to start up
         alive_url = "http://{host}:{port}/_alive_{token}".format(